替换UID模块
