+++ agent/go-service/creditshopping/recognition_results.go
@@ -78,3 +78,23 @@ func sortOCRNameHits(hits []ocrNameHit) {
 		return hits[i].Box[0] < hits[j].Box[0]
 	})
 }
+
+func bestOCRText(detail *maa.RecognitionDetail) string {
+	if detail == nil || detail.Results == nil {
+		return ""
+	}
+	if detail.Results.Best != nil {
+		if o, ok := detail.Results.Best.AsOCR(); ok {
+			return strings.TrimSpace(o.Text)
+		}
+	}
+	for _, r := range detail.Results.Filtered {
+		if r == nil {
+			continue
+		}
+		if o, ok := r.AsOCR(); ok {
+			return strings.TrimSpace(o.Text)
+		}
+	}
+	return ""
+}
